refactor: move section tests into sections/tests/ package
Each section now has its own test file next to the code it tests:
- sections/tests/test_personal_bests.py
- sections/tests/test_profile.py
- sections/tests/test_training_history.py

test_context.py retains only TestBuildCoachContext and the golden-output test.
